LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-duplicate-importmap-expected.txt


FAIL Duplicate bare specifiers via import map are deduplicated and applied. assert_equals: Duplicate bare specifiers via import map should be deduplicated. expected 1 but got 0
